requirements.txt
discord.py==2.6.4
path==17.1.1